src/backend/requirements.txt
# ============================================================
# Money Hunter Backend — Requirements
# Maintainer: Young-cheol (Backend Engineer)
# ============================================================

# --- Web Framework ---
fastapi>=0.115.0
uvicorn[standard]>=0.34.0

# --- Browser Automation ---
playwright>=1.49.0

# --- Scheduling ---
apscheduler>=3.10.0

# --- Configuration ---
pydantic-settings>=2.7.0
python-dotenv>=1.0.0

# --- Data Processing ---
pandas>=2.2.0
beautifulsoup4>=4.12.0

# --- Database ---
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.30.0

# --- HTTP Client ---
httpx>=0.28.0
